Add unit tests for Unity exporter template
